requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
flask==2.3.2
joblib==1.3.1
pyyaml==6.0
pytest==7.4.0